Removed mage inheriting from EnemyStateManager
Do not delete. Will merge in this branch once it is possible.

Since mage implementation doesn't actually support EnemyStateManager, a bunch of errors will be thrown when mage inherits from it. But considering that certain scripts might be calling the EnemyStateManager script on mage, this change won't be implemented just yet.